% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{countAmplicons}
\alias{countAmplicons}
\title{Count expected amplicons}
\usage{
countAmplicons(
  in.con,
  index.key,
  amplicons,
  line.buffer = 5e+06,
  max.lines = NULL,
  nthreads = 1
)
}
\arguments{
\item{in.con}{a file connections, for example gzfile}

\item{index.key}{a data.table with Plate_ID, Sample_Well, Sample_ID, index, and index2  ... where index is i7 and index2 is i5}

\item{amplicons}{a list of named expected sequences for read 1 amplicons}

\item{line.buffer}{an integer specifiying the number of lines to read at once}

\item{max.lines}{an integer specifiying the maximum total number of lines to read for debugging (default=NULL)}

\item{nthreads}{an integer specifiying the maximum number of threads (default=1)}
}
\value{
a list containing \code{count.tables} for each expected amplicon and \code{amp.match.summary} the number of reads across the experiment matching each expected amplicon
}
\description{
Count the expected amplicons for the given indices
}
