% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{errorCorrectIdxAndCountAmplicons}
\alias{errorCorrectIdxAndCountAmplicons}
\title{error correct index reads and assign amplicon counts}
\usage{
errorCorrectIdxAndCountAmplicons(rid, count.table, ind1, ind2)
}
\arguments{
\item{rid}{indices of sequences matching an expected amplicon}

\item{count.table}{table per amplicon to hold amplicon counts per expected index}

\item{ind1}{index1 sequences}

\item{ind2}{index2 sequences}
}
\value{
count.table the set of all unique hamming distance 1 sequences from a given sequence
}
\description{
error correct index reads and assign amplicon counts
}
