% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{initAmpliconCountTables}
\alias{initAmpliconCountTables}
\title{initialize amplicon count table}
\usage{
initAmpliconCountTables(index.key, amplicons)
}
\arguments{
\item{index.key}{index.key from generateExpectedIndices}

\item{amplicons, }{a list containing names and expected sequences of amplicons}
}
\value{
table per amplicon to hold amplicon counts per expected index
}
\description{
initialize amplicon count table
}
