% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{make_hamming1_sequences}
\alias{make_hamming1_sequences}
\title{make all hamming 1 distant sequences for a given sequence}
\usage{
make_hamming1_sequences(x)
}
\arguments{
\item{x}{a sequences}
}
\value{
the set of all unique hamming distance 1 sequences from a given sequence
}
\description{
make all hamming 1 distant sequences for a given sequence
}
