% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreFunctions.R
\name{revcomp2}
\alias{revcomp2}
\title{Reverse complement a sequence}
\usage{
revcomp2(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
reverse complement of x
}
\description{
\code{revcomp} returns the reverse complement of a sequence
}
